Split controllers into separate packages.
